Expanded small caps section (complete through n)
The policy is to include small caps for all characters for which there are both lowercase and uppercase glyphs. Some PUA characters are included, but unsure whether to include, e.g. precomposed characters in the PUA. There are a great many of these, but omitting them could cause problems, and each costs only a few bytes, since they are constructed from components.